pyproject.toml
[tool.poetry]
name = "cds-302-project"
version = "0.1.0"
description = ""
authors = ["Tyson Johnson <tjohns94@gmu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pathlib = "^1.0.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
logging = "^0.4.9.6"
sqlalchemy-utils = "^0.41.2"
argparse = "^1.4.0"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
